+++ crest/Assets/Crest/Crest-Examples/Examples/Scripts/CustomPassForCamera.cs
@@ -9,7 +9,6 @@ namespace Crest.Examples
     using UnityEngine.Events;
     using UnityEngine.Rendering;
 
-    [ExecuteAlways]
     public abstract class CustomPassForCameraBase : MonoBehaviour
     {
         // Use int to support other RPs. We could use a custom enum to map to each RP in the future.
@@ -92,6 +91,13 @@ namespace Crest.Examples
 
         protected abstract void Execute(CommandBuffer buffer, Camera camera, RenderTextureDescriptor descriptor);
         protected abstract void Clear(CommandBuffer buffer, Camera camera);
+
+#if UNITY_EDITOR
+        void OnValidate()
+        {
+            runInEditMode = !UnityEditor.BuildPipeline.isBuildingPlayer;
+        }
+#endif
     }
 
     public class CustomPassForCamera : CustomPassForCameraBase