+++ crest/Assets/Crest/Crest/Scripts/OceanChunkRenderer.cs
@@ -14,7 +14,6 @@ namespace Crest
     /// <summary>
     /// Sets shader parameters for each geometry tile/chunk.
     /// </summary>
-    [ExecuteAlways]
     [AddComponentMenu(Internal.Constants.MENU_PREFIX_INTERNAL + "Ocean Chunk Renderer")]
     public class OceanChunkRenderer : MonoBehaviour
     {
@@ -200,6 +199,12 @@ namespace Crest
             RenderPipelineManager.beginCameraRendering += BeginCameraRendering;
         }
 
+#if UNITY_EDITOR
+        void OnValidate()
+        {
+            runInEditMode = !UnityEditor.BuildPipeline.isBuildingPlayer;
+        }
+
         private void OnDrawGizmos()
         {
             if (_drawRenderBounds)
@@ -207,6 +212,7 @@ namespace Crest
                 Rend.bounds.GizmosDraw();
             }
         }
+#endif
     }
 
     public static class BoundsHelper